Implemented weighted random outputs for fluid_spread recipe
This is an optional feature. Recipes can also just request a single output block.
If multiple blocks are specified, a weight decides how likely each of them is.
A higher value makes the block more likely.
How likely depends on the total of all weights. Outputs which don't provide a weight default to a weight of 1.