Fix crash when running server side
Turns out that the event RecipesUpdatedEvent is only for clients.
There is no equivalent event for servers sadly.

Instead, the cache was modified to automatically populate if you ask for recipes without having populated it first.
On the server side, OnDatapackSyncEvent can be used to detect when datapacks are reloaded which we use to invalidate the recipe cache.