Update to MC 1.19.2
For such a small mod, surprisingly so much changed...

Forge for 1.19 introduced the FluidInteractionRegistry. It adds a way to create new fluid interactions like this mod does.
However, the way forge does it is significantly more limited. You may only register the interactions at startup, and there is no way to override an existing interaction.
Due to its design, it was not possible to use it to implement this mod. While the HasFluidInteraction and FluidInteraction interfaces are flexible enough to work with us, the possible interactions are stored in a map by fluid type.
So we would need to know what fluid types will be used in the recipes ahead of time which obviously doesn't work.

Our recipes have been set to take priority over the FluidInteractionRegistry, but if we do not have a recipe then it will be handled by the registry.