Add JEI recipe support to display the recipe type
Players with JEI installed can now lookup the recipe on any output block and see the setup needed to create it.
The visual layout of the recipe is changed depending on the recipe.
Recipes involving two liquids show that they should interact.
Recipes involving a block show that the liquid must flow into the block.
Recipes which are stone-like (falling recipes) show that the liquid must flow down.

The probability of each output is also displayed if there are multiple outputs.