Add forEachMatchingRegisteredBlock to BlockMatcher
This iterates through all registered block types, finding ones that match the BlockMatcher.
We can do this since the list of blocks in the game is fixed and we only allow matching on
block id and taglists.
If we allow matching on block state or data, then this method would be inaccurate.