Add rockgen:dripstone_draining recipe type
This recreates the Mud to Clay conversion process as a recipe type.
The recipe can configure:

- The block or blocks to convert (supports taglists)
- The particle effect while conversion is in progress
- The probability of the conversion
- The result block
- Whether it works in the nether (or other hot dimensions)

The output block supports weighted random selection. It is configured the same way as the fluid_spread recipe.
Probability is a percentage of how likely any random tick is to trigger the conversion.
The particle effect can be any simple particle effect (one that takes no parameters). It is displayed on the tip of the dripstone stalactite. If omitted, it will just show water drips.