Added JEI support for new dripstone_draining recipe type
It shows a small diagram of how to set up for draining.
The displayed root block is stone, but this might be a bit unclear as this can really be any block that supports pointed dripstone. There is a tooltip on that to help clear that up.

There is an icon that displays (with tooltip) for recipes which cannot be used in the nether.
The probability is displayed. Would have loved to say how long on average this happens, but no idea what I would need to do for that as the probability is based on random ticks.