Allow recipes to prevent the events from changing the result block
Some mods (such as Create) use the FluidPlaceBlockEvent to set up their own block
generation. This event happens after our recipes would take effect making it
impossible to add recipes which involve their fluids in any way.

This change adds the "allow_override" flag that recipes can set.
When enabled, the block set in FluidPlaceBlockEvent is ignored.
By default the value is true so we do not alter the behaviour of the system.