Fabric: Port regular functionality to Fabric
Mixins are now applied as expected.
The LavaFluidMixin spreadTo method did need to have its access raised to public. I assume there is an access transformer in Fabric that raises that from its usual protected access.
It would be nice to work with this without needing to raise access in forge too.

The recipes are now loaded correctly and the caches reloaded when required.