Fabric: Port JEI integration
This is a duplicate of the Forge integration.
It does not use common code because of slight differences between the two platforms.

There is a visual difference between the two platforms with the fluid stack rendering.
In Forge, JEI shows the fluid in the entire slot. In Fabric, JEI shows a thin line at the bottom of the slot.